Improve particle weights in initial condition
Before, all particles in a cell had the same integer weight, which could lead to
a somewhat too high or too low density. This is now fixed. The weights in a cell
can now be either w or w + particle_min_weight, with w being a multiple of
particle_min_weight.